Extract PDF page stuff to its own file
Now that the page content stuff is decoupled from saving each page, we
pull it out of the `import-pdf` app and into the more general
`fabula-ultima-pdf` parser. With the logic here, we can use it for the
test as well, instead of needing to redefine it here.

We also fix the typo in the name of the test file from
`src/pdf/parsers/fabula-ultma-pdf.test.ts` to
`src/pdf/parsers/fabula-ultima-pdf.test.ts`.
